forecast-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
